requirements.txt
asgiref==3.11.0
Django==6.0
django-restframework==0.0.1
pillow==12.0.0
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
